Fixed collecting page crashing
Collecting page no longer crashes if the player's priority decks contains a deck not yet in the UPDATEcolorstcgdeckdata.js dictionary (the player has not updated the file after a new release, and has set a new release deck in high priority). If data cannot be found for the deck, the deck is still displayed, just without a header color. It also appears in a blank column for the all iso decks section.